backend/pyproject.toml
[project]
name = "alphaquant-backend"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sse-starlette>=2.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "langgraph>=0.2.60",
    "langchain-core>=0.3.0",
    "pyyaml>=6.0",
    "beautifulsoup4>=4.12",
    # --- Auth + persistence (Phase 2) ---
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "bcrypt>=4.1",
    "python-jose[cryptography]>=3.3",
    "authlib>=1.3",          # Google OAuth
    "itsdangerous>=2.2",     # signed magic-link tokens
    "email-validator>=2.1",  # pydantic EmailStr support
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.24", "pytest-httpx>=0.34"]
